revert to CRS namespace/class for now
